+++ openpype/hosts/hiero/api/events.py
@@ -1,12 +1,12 @@
 import os
 import hiero.core.events
 from openpype.api import Logger
-from openpype.pipeline import register_event_callback
 from .lib import (
     sync_avalon_data_to_workfile,
     launch_workfiles_app,
     selection_changed_timeline,
-    before_project_save
+    before_project_save,
+    register_event_callback
 )
 from .tags import add_tags_to_workfile
 from .menu import update_menu_task_label
