+++ openpype/pipeline/__init__.py
@@ -1,10 +1,5 @@
 from .lib import attribute_definitions
 
-from .events import (
-    emit_event,
-    register_event_callback
-)
-
 from .create import (
     BaseCreator,
     Creator,
@@ -21,9 +16,6 @@ from .publish import (
 
 __all__ = (
     "attribute_definitions",
-
-    "emit_event",
-    "register_event_callback",
 
     "BaseCreator",
     "Creator",
